updated routing so works with fleek
