+++ poa-app/src/pages/dashboard/index.js
@@ -1,83 +1,394 @@
-import React from "react";
-import { Box, Flex, VStack } from "@chakra-ui/react";
-import { motion } from "framer-motion";
-import Layout from "../../components/Layout";
-import AboutPreview from "@/components/Dashboard/AboutPreview";
-import YourPerpetualOrganizations from "@/components/Dashboard/YourPerpetualOrganizations";
-import PoaPreview from "@/components/Dashboard/PoaPreview";
+import React, { useState, useEffect } from 'react';
+import {
+  Box,
+  Button,
+  VStack,
+  Grid,
+  GridItem,
+  Text,
+  IconButton,
+  HStack,
+  keyframes,
+  usePrefersReducedMotion,
+  chakra,
+  Image,
+  Progress,
+  Spacer,
+    Badge,
+  
+} from '@chakra-ui/react';
+import { SettingsIcon } from '@chakra-ui/icons';
+// import AccountSettingsModal from '@/components/userPage/AccountSettingsModal';
+
+import { useWeb3Context } from '@/context/web3Context';
+import { useGraphContext } from '@/context/graphContext';
+
+
+// import DeployMenu from "@/components/userPage/DeployMenu";
+// import MintMenu from "@/components/userPage/MintMenu";
+// import DataMenu from "@/components/userPage/DataMenu";
+
+
+import { useSpring, animated } from 'react-spring';
+
+
+import Link2 from 'next/link';
+import { set } from 'lodash';
+import OngoingPolls from '@/components/userPage/OngoingPolls';
+import { useRouter } from 'next/router';
+import Navbar from "@/templateComponents/studentOrgDAO/NavBar";
 
 
 
-// Convert Chakra UI components to motion components
-const MotionBox = motion(Box);
-const MotionFlex = motion(Flex);
 
-const DashboardPage = () => {
-  // Animation variants
-  const containerVariants = {
-    hidden: { opacity: 0 },
-    visible: {
-      opacity: 1,
-      transition: {
-        staggerChildren: 0.2, // Stagger the animations of children
-      },
-    },
+
+const UserDashboard= () => {
+    const { userData, setLoaded} = useGraphContext();
+    const router = useRouter();
+    const { userDAO } = router.query;
+    useEffect(() => {
+        setLoaded(userDAO);
+      }, [userDAO]);
+  
+
+    // const {setLeaderboardLoaded,userPercentage} = useLeaderboard();
+
+    // const {loadOngoingKubidInitial, kubidOngoingProposals} = useGraphVotingContext();
+    
+
+    
+    
+  const prefersReducedMotion = usePrefersReducedMotion();
+  const [countFinished, setCountFinished] = useState(false);
+
+  const[upgradeAvailable, setUpgradeAvailable] = useState(false);
+
+  const [showDevMenu, setShowDevMenu] = useState(false);
+
+  const toggleDevMenu = () => {
+    setShowDevMenu(prevShowDevMenu => !prevShowDevMenu);
+};
+
+  const [isSettingsModalOpen, setSettingsModalOpen] = useState(false);
+
+  const openSettingsModal = () => setSettingsModalOpen(true);
+  const closeSettingsModal = () => setSettingsModalOpen(false);
+
+  const [claimedTasks, setClaimedTasks] = useState([]);
+    const [notLoaded, setNotLoaded] = useState(true);
+  const [reccomendedTasks, setReccomendedTasks] = useState([]);
+
+
+    const { web3, account,KUBIXbalance, hasExecNFT} = useWeb3Context();
+    // const {userDetails, fetchUserDetails, findUserInProgressTasks, projects, findRandomTasks} = useDataBaseContext();
+
+
+    // useEffect(() => {
+        
+    //     async function fetch(){
+    //         await fetchUserDetails(web3,account);
+    //         await setLeaderboardLoaded(true);
+    //         await loadOngoingKubidInitial();
+            
+            
+    //     }
+    //     if(web3 && account&& notLoaded){
+    //         setNotLoaded(false)
+
+    //         fetch();
+    //     }
+        
+    //   }, [web3, account]);
+
+    //   useEffect(() => {
+    //         async function fetch(){
+    //             if (userDetails && projects){
+    //                 console.log(userDetails.username)
+    //                 console.log(projects)
+    //                 const wait =await findUserInProgressTasks(projects , userDetails.username)
+                   
+    //                 if(wait.length === 0){
+    //                     console.log("ready");
+    //                     const random = await findRandomTasks(projects, 3);
+    //                     setReccomendedTasks(random);
+    //                     console.log(random);
+    //                 }
+
+    //                 setClaimedTasks(wait);
+    //                 console.log(wait);
+
+    //             }
+    //             else{
+    //                 console.log("not ready");
+        
+    //             }
+                
+
+    //         }
+    //         fetch();
+    //     }, [userDetails, projects]);
+
+        
+
+
+  const glassLayerStyle = {
+    position: 'absolute',
+    height: '100%',
+    width: '100%',
+    zIndex: -1,
+    borderRadius: 'inherit',
+    backdropFilter: 'blur(20px)',
+    backgroundColor: 'rgba(0, 0, 0, .7)',
   };
 
-  const itemVariants = {
-    hidden: { y: 20, opacity: 0 },
-    visible: {
-      y: 0,
-      opacity: 1,
-      transition: { duration: 0.6 },
-    },
+  const glowAnimation = keyframes`
+    from { text-shadow: 0 0 0px white; }
+    to { text-shadow: 0 0 20px gold;  }
+  `;
+
+
+  const determineTier = (kubixBalance) => {
+    if (kubixBalance >= 2500) return "Diamond";
+    else if (kubixBalance >= 1000) return "Ruby";
+    else if (kubixBalance >= 500) return "Gold";
+    else if (kubixBalance >= 250) return "Silver";
+    else if (kubixBalance >= 100) return "Bronze";
+    else return "New"; 
   };
+
+  const tierInfo = {
+    Diamond: {
+        nextTier: "Double Diamond",
+        nextTierKUBIX: "5000",
+        nextTierReward: "Flex",
+        image: "/images/diamondMember.png"
+    },
+    Ruby: {
+        nextTier: "Diamond",
+        nextTierKUBIX: 2500,
+        nextTierReward: "Choice of Any Item",
+        image: "/images/rubyMember.png"
+    },
+    Gold: {
+        nextTier: "Ruby",
+        nextTierKUBIX: 1000,
+        nextTierReward: "Choice of Shirt, Hat, or Pullover",
+        image: "/images/goldMember.png"
+    },
+    Silver: {
+        nextTier: "Gold",
+        nextTierKUBIX: 500,
+        nextTierReward: "Choice of Shirt or Hat",
+        image: "/images/silverMember.png"
+    },
+    Bronze: {
+        nextTier: "Silver",
+        nextTierKUBIX: 250,
+        nextTierReward: "Choice of Shirt",
+        image: "/images/bronzeMember.png"
+    },
+    New: {
+        nextTier: "Bronze",
+        nextTierKUBIX: 100,
+        nextTierReward: "T-shirt and Trip Eligibility",
+        image:"/images/newMember.png",
+    },
+    };
+
+
+    const [userInfo, setUserInfo] = useState({});
+
+    useEffect(() => {
+        console.log("checking user data", userData)
+        if(userData){
+            console.log(userData);
+            let userInfo = {
+                username: userData.id,
+                ptBalance: Number(userData.ptTokenBalance),
+                memberStatus: userData.memberType?.memberTypeName,
+                accountAddress: userData.id,
+            };
+            setUserInfo(userInfo);
+        }
+    }, [userData]);
+
+
+  const animatedPT = useSpring({ 
+    pt: userInfo.ptBalance, 
+    from: { pt: 0 },
+    config: { duration: 1700 },
+    onRest: () => setCountFinished(true),
+  });
+  
+
+  const animationProps = prefersReducedMotion
+    ? {}
+    : {
+        animation: `${glowAnimation} alternate 2.1s ease-in-out`,
+      };
+
 
   return (
-    <Layout>
-      <MotionBox
-        pt="4rem"
-        w="full"
-        overflowY="auto"
-        variants={containerVariants}
-        initial="hidden"
-        animate="visible"
+    <>
+    <Navbar />
+    
+    <Box p={4}>
+      <Grid
+        color="white"
+        templateAreas={[
+          `'welcome welcome' 'userinfo tierinfo' 'userinfo tierinfo'`,
+          `'welcome welcome' 'userinfo tierinfo' 'userinfo tierinfo'`
+        ]}
+        templateColumns="repeat(2, 1fr)"
+        gap={4}
       >
-        <VStack mt= "6" spacing={6} w="full" h="full" justify="center" >
 
-
-          <MotionFlex
-            w="full"
-            flex="1"
-            direction={{ base: "column", md: "row" }}
-            p={3}
-            variants={itemVariants}
-          >
-            <MotionBox
-              w={{ base: "full", md: "50%" }}
-              h="full"
-              p={3}
-              variants={itemVariants}
-            >
-              <PoaPreview />
-            </MotionBox>
-            <MotionBox
-              w={{ base: "full", md: "50%" }}
-              h="full"
-              p={3}
-              variants={itemVariants}
-            >
-              <YourPerpetualOrganizations />
-            </MotionBox>
+        <GridItem area={'userinfo'}>
+        <Box
+            w="100%"
+            borderRadius="2xl"
+            bg="transparent"
+            boxShadow="lg"
+            position="relative"
+            zIndex={2}
             
-          </MotionFlex>
+          >
+        <div style={glassLayerStyle} />
+          <VStack position="relative" borderTopRadius="2xl" align="flex-start">
+          <div style={glassLayerStyle} />
+            <Text pl={6} letterSpacing="-1%" mt={2} fontSize="4xl" id="kubix-earned" fontWeight="bold">
+                Tokens Earned {' '}
+                {countFinished ? (
+                <chakra.span {...animationProps}>{userInfo.ptBalance}</chakra.span>
+
+            ) : (
+                <animated.span>
+                    {animatedPT.pt.to(pt => pt.toFixed(0))}
+                </animated.span>)}
+                </Text>
+
+            {/* <Text pl={6} pb={4} fontSize="lg">This makes you top {userInfo.percentage}% of Contributors</Text> */}
+          </VStack>
+            <VStack p={0} pt={4} align="center" >
+                <Text fontSize="3xl" fontWeight="bold">{userInfo.memberStatus} Member</Text>
+                <Spacer />
+            </VStack>
+
+        </Box>
+        </GridItem>
+        <GridItem area={'tierinfo'}>
+        <Box
+            w="100%"
+            borderRadius="2xl"
+            bg="transparent"
+            boxShadow="lg"
+            position="relative"
+            zIndex={2}
+            
+          >
+        <div style={glassLayerStyle} />
+        <HStack  pt={4} pb= {4} position="relative" borderTopRadius="2xl" >
+        <div style={glassLayerStyle} />
+          <Text  pl={4}  fontSize="3xl" fontWeight="extrabold">{userInfo.username} </Text>
+          <Text pt={2} pl={2} fontSize="lg" > {userInfo.memberStatus}</Text>
+        </HStack>
+          {/* <IconButton
+          icon={<SettingsIcon />}
+          isRound={true}
+          size="md"
+          aria-label="Settings"
+          onClick={openSettingsModal}
+          alignSelf="start"
+          justifySelf="end"
+          position="absolute"
+          top="8%"
+          right="4%"
+          color="black"
+        /> */}
+        
+        {/* <AccountSettingsModal
+        isOpen={isSettingsModalOpen}
+        onClose={closeSettingsModal}
+      /> */}
+        <HStack pb={4} pt={2}  spacing="27%">
+            <VStack align={'flex-start'} ml="6%" spacing={1}>
+                {/* <Text   fontWeight="bold" fontSize="md">Semester KUBIX: {userInfo.semesterKubix}</Text>
+                <Text  fontWeight="bold" fontSize="md">Year KUBIX: {userInfo.yearKubix}</Text> */}
+                <Text  fontWeight="bold"  fontSize="md">Tasks Completed: {userInfo.tasksCompleted}</Text>
+            </VStack>
+            <VStack align={'center'} spacing={2}>
+                {hasExecNFT && (
+                <Button colorScheme="green" onClick={toggleDevMenu} size="sm">Developer Menu</Button>
+                )}
+                {!hasExecNFT &&(<Button colorScheme="red" size="sm">Become Developer</Button>)}
+                {showDevMenu && (
+                <>
+                    <DeployMenu />
+                    <MintMenu />
+                    <DataMenu />
+                </>
+            )}
+               
+            </VStack>
+            </HStack>
+        </Box>
+        <Box w="100%" pt={4} borderRadius="2xl" bg="transparent" boxShadow="lg" position="relative" zIndex={2} >
+            <div style={glassLayerStyle} />
+
+            <VStack pb={2} align="flex-start" position="relative" borderTopRadius="2xl">
+                <div style={glassLayerStyle} />
+                <Text pl={6} fontWeight="bold" fontSize="2xl">
+                    {claimedTasks && claimedTasks.length > 0 ? 'Claimed Tasks' : 'Recommended Tasks'}
+                </Text>
+            </VStack>
+
+            {/* <HStack spacing="3.5%" pb={2} ml={4} mr={4} pt={4}>
+                {((claimedTasks && claimedTasks.length > 0) ? claimedTasks : reccomendedTasks)?.slice(0, 3).map((task) => (
+                    <Box key={task.id} w="31%" _hover={{ boxShadow: "md", transform: "scale(1.07)"}} p={4} borderRadius="2xl" overflow="hidden" bg="black">
+                        <Link2 href={`/tasks/?task=${task.id}&projectId=${task.projectId}`}>
+                            <VStack textColor="white" align="stretch" spacing={3}>
+                                <Text fontSize="md" lineHeight="99%" fontWeight="extrabold">
+                                    {task.name.length > 57 ? `${task.name.substring(0, 57)}...` : task.name}
+                                </Text>
+                                <HStack justify="space-between">
+                                    <Badge colorScheme="yellow">{task.difficulty}</Badge>
+                                    <Text fontWeight="bold">KUBIX {task.kubixPayout}</Text>
+                                </HStack>
+                            </VStack>
+                        </Link2>
+                    </Box>
+                ))}
+            </HStack> */}
+
+        </Box>
+
+            <Box w="100%"
+            pt={8}
+            borderRadius="2xl"
+            bg="transparent"
+            boxShadow="lg"
+            position="relative"
+            zIndex={2}>
+        <div style={glassLayerStyle} />
+
+        <VStack pb={2}  align="flex-start" position="relative" borderTopRadius="2xl">
+        <div style={glassLayerStyle} />
+            <Text pl={6} fontWeight="bold" fontSize="2xl" >Ongoing Polls {' '}</Text>
+            
+
         </VStack>
-        <MotionBox w="full" flex="1" p={3} variants={itemVariants}>
-            <AboutPreview />
-          </MotionBox>
-      </MotionBox>
-    </Layout>
+
+        {/* <OngoingPolls  OngoingPolls={kubidOngoingProposals}/> */}
+
+            
+        </Box>
+
+        </GridItem>
+        
+      </Grid>
+    </Box>
+    </>
   );
 };
 
-export default DashboardPage;
+export default UserDashboard;
